Feat(user): 회원 탈퇴 기능 및 권한 부여 테스트 추가 - 사용자 본인만 탈퇴할 수 있도록 이메일 검사 추가 - 권한 불일치 시 /main으로 리다이렉트 - 회원 탈퇴 시 세션 초기화 및 /login으로 이동 - 레벨에 따른 권한 부여 테스트 추가 - 사용자 본인인 경우 즉시 권한 갱신
